requirements.txt
decorator==4.0.11
matplotlib==2.0.2
numpy==1.13.1
pandas==0.20.2
Pillow==3.4.2
protobuf==3.3.0
pygpu==0.6.9
pyparsing==2.1.4
scipy==0.19.1
seaborn==0.7.1
simplejson==3.11.1
tensorboard==1.0.0a6
tensorflow==1.2.1
tqdm==4.11.2
virtualenv==15.1.0
pytorch==0.1.12
progressbar2==3.18.1
termcolor==1.1.0
